[Metricbeat] Separate vpc metricset to vpn, transitgateway and natgateway metricsets (#16892)
* Separate vpc metricset to vpn, transitgateway and natgateway metricsets
* add dashboard and screenshot 

(cherry picked from commit 7a7c2afa3ce753e517f053b8665071ac056943ea)
