+++ metricbeat/docs/fields.asciidoc
@@ -1973,6 +1973,12 @@ type: integer
 
 
 [float]
+=== natgateway
+
+`natgateway` contains the metrics from Cloudwatch to track usage of NAT gateway related resources.
+
+
+[float]
 === rds
 
 `rds` contains the metrics that were scraped from AWS CloudWatch which contains monitoring metrics sent by AWS RDS.
@@ -3083,15 +3089,21 @@ type: keyword
 --
 
 [float]
+=== transitgateway
+
+`transitgateway` contains the metrics from Cloudwatch to track usage of transit gateway related resources.
+
+
+[float]
 === usage
 
 `usage` contains the metrics from Cloudwatch to track usage of some AWS resources.
 
 
 [float]
-=== vpc
+=== vpn
 
-`vpc` contains the metrics from Cloudwatch to track usage of VPC related resources.
+`vpn` contains the metrics from Cloudwatch to track usage of VPN related resources.
 
 
 [[exported-fields-azure]]
