+++ metricbeat/docs/modules/aws.asciidoc
@@ -35,8 +35,8 @@ image::./images/metricbeat-aws-overview.png[]
 == Metricsets
 
 Currently, we have `billing`, `cloudwatch`, `dynamodb`, `ebs`, `ec2`, `elb`,
-`lambda`, `rds`, `s3_daily_storage`, `s3_request`, `sns`, `sqs` and `usage`
-metricset in `aws` module.
+`lambda`, `natgateway`, `rds`, `s3_daily_storage`, `s3_request`, `sns`, `sqs`,
+`transitgateway`, `usage` and `vpn` metricset in `aws` module.
 
 Collecting `tags` for `ec2`, `rds`, `cloudwatch`, and metricset created based on
 `cloudwatch` using light module is supported.
@@ -125,6 +125,13 @@ The lambda metricset comes with a predefined dashboard:
 image::./images/metricbeat-aws-lambda-overview.png[]
 
 [float]
+=== `natgateway`
+CloudWatch collects information from NAT gateways and creates readable, near real-time metrics.
+This metricset enables users to collect these metrics from CloudWatch to monitor and troubleshoot
+their NAT gateway. NAT gateway metric data is provided at 1-minute intervals and therefore,
+`period` for `natgateway` metricset is recommended to be `1m` or multiples of `1m`.
+
+[float]
 === `rds`
 `period` for `rds` metricset is recommended to be `60s` or multiples of `60s` because Amazon RDS sends metrics and
 dimensions to Amazon CloudWatch every minute.
@@ -163,6 +170,13 @@ The sqs metricset comes with a predefined dashboard:
 image::./images/metricbeat-aws-sqs-overview.png[]
 
 [float]
+=== `transitgateway`
+Amazon VPC reports metrics to CloudWatch only when requests are flowing through
+the transit gateway. If there are requests flowing through the transit gateway,
+Amazon VPC measures and sends its metrics in 60-second intervals. `period` for
+`transitgateway` metricset is recommended to be `1m` or multiples of `1m`.
+
+[float]
 === `usage`
 CloudWatch collects metrics that track the usage of some AWS resources. These
 metrics correspond to AWS service quotas. Tracking these metrics can help
@@ -173,6 +187,12 @@ configuration for usage metricset is set to `1m`.
 The usage metricset comes with a predefined dashboard:
 
 image::./images/metricbeat-aws-usage-overview.png[]
+
+[float]
+=== `vpn`
+CloudWatch collects and processes raw data from the VPN service into readable, near
+real-time metrics for users to better understand the performance of their web
+applications and services.
 
 [float]
 == AWS API requests count per metricset
@@ -305,7 +325,12 @@ metricbeat.modules:
     - key: "dept"
       value: "eng"
   metricsets:
+    - elb
+    - natgateway
     - rds
+    - transitgateway
+    - usage
+    - vpn
 ----
 
 [float]
@@ -327,6 +352,8 @@ The following metricsets are available:
 
 * <<metricbeat-metricset-aws-lambda,lambda>>
 
+* <<metricbeat-metricset-aws-natgateway,natgateway>>
+
 * <<metricbeat-metricset-aws-rds,rds>>
 
 * <<metricbeat-metricset-aws-s3_daily_storage,s3_daily_storage>>
@@ -337,9 +364,11 @@ The following metricsets are available:
 
 * <<metricbeat-metricset-aws-sqs,sqs>>
 
+* <<metricbeat-metricset-aws-transitgateway,transitgateway>>
+
 * <<metricbeat-metricset-aws-usage,usage>>
 
-* <<metricbeat-metricset-aws-vpc,vpc>>
+* <<metricbeat-metricset-aws-vpn,vpn>>
 
 include::aws/billing.asciidoc[]
 
@@ -355,6 +384,8 @@ include::aws/elb.asciidoc[]
 
 include::aws/lambda.asciidoc[]
 
+include::aws/natgateway.asciidoc[]
+
 include::aws/rds.asciidoc[]
 
 include::aws/s3_daily_storage.asciidoc[]
@@ -365,7 +396,9 @@ include::aws/sns.asciidoc[]
 
 include::aws/sqs.asciidoc[]
 
+include::aws/transitgateway.asciidoc[]
+
 include::aws/usage.asciidoc[]
 
-include::aws/vpc.asciidoc[]
+include::aws/vpn.asciidoc[]
