+++ metricbeat/docs/modules/aws/transitgateway.asciidoc
@@ -1,0 +1,24 @@
+////
+This file is generated! See scripts/mage/docs_collector.go
+////
+
+[[metricbeat-metricset-aws-transitgateway]]
+=== aws transitgateway metricset
+
+beta[]
+
+include::../../../../x-pack/metricbeat/module/aws/transitgateway/_meta/docs.asciidoc[]
+
+This is a default metricset. If the host module is unconfigured, this metricset is enabled by default.
+
+==== Fields
+
+For a description of each field in the metricset, see the
+<<exported-fields-aws,exported fields>> section.
+
+Here is an example document generated by this metricset:
+
+[source,json]
+----
+include::../../../../x-pack/metricbeat/module/aws/transitgateway/_meta/data.json[]
+----
