+++ metricbeat/docs/modules/aws/vpn.asciidoc
@@ -2,12 +2,12 @@
 This file is generated! See scripts/mage/docs_collector.go
 ////
 
-[[metricbeat-metricset-aws-vpc]]
-=== aws vpc metricset
+[[metricbeat-metricset-aws-vpn]]
+=== aws vpn metricset
 
 beta[]
 
-include::../../../../x-pack/metricbeat/module/aws/vpc/_meta/docs.asciidoc[]
+include::../../../../x-pack/metricbeat/module/aws/vpn/_meta/docs.asciidoc[]
 
 This is a default metricset. If the host module is unconfigured, this metricset is enabled by default.
 
@@ -20,5 +20,5 @@ Here is an example document generated by this metricset:
 
 [source,json]
 ----
-include::../../../../x-pack/metricbeat/module/aws/vpc/_meta/data.json[]
+include::../../../../x-pack/metricbeat/module/aws/vpn/_meta/data.json[]
 ----
