+++ x-pack/metricbeat/module/aws/_meta/config.reference.yml
@@ -45,4 +45,9 @@
     - key: "dept"
       value: "eng"
   metricsets:
+    - elb
+    - natgateway
     - rds
+    - transitgateway
+    - usage
+    - vpn
