+++ x-pack/metricbeat/module/aws/module.yml
@@ -7,4 +7,6 @@ metricsets:
   - sns
   - lambda
   - dynamodb
-  - vpc
+  - vpn
+  - transitgateway
+  - natgateway
