+++ x-pack/metricbeat/module/aws/transitgateway/_meta/fields.yml
@@ -1,0 +1,6 @@
+- name: transitgateway
+  type: group
+  description: >
+    `transitgateway` contains the metrics from Cloudwatch to track usage of transit gateway related resources.
+  release: beta
+  fields:
