+++ x-pack/metricbeat/module/aws/transitgateway/manifest.yml
@@ -1,0 +1,15 @@
+default: true
+input:
+  module: aws
+  metricset: cloudwatch
+  defaults:
+    metrics:
+      - namespace: AWS/TransitGateway
+        statistic: ["Sum"]
+        name:
+          - BytesIn
+          - BytesOut
+          - PacketsIn
+          - PacketsOut
+          - PacketDropCountBlackhole
+          - PacketDropCountNoRoute
