+++ x-pack/metricbeat/module/aws/vpn/manifest.yml
@@ -1,0 +1,15 @@
+default: true
+input:
+  module: aws
+  metricset: cloudwatch
+  defaults:
+    metrics:
+      - namespace: AWS/VPN
+        statistic: ["Average"]
+        name:
+          - TunnelState
+      - namespace: AWS/VPN
+        statistic: ["Sum"]
+        name:
+          - TunnelDataIn
+          - TunnelDataOut
