+++ x-pack/metricbeat/modules.d/aws.yml.disabled
@@ -5,7 +5,11 @@
   period: 1m
   metricsets:
     - elb
+    - natgateway
+    - rds
+    - transitgateway
     - usage
+    - vpn
 - module: aws
   period: 5m
   metricsets:
